django_informer/requirements.txt
Django>=4.2.0
torch>=1.9.0
numpy>=1.19.0
PyYAML>=6.0
easydict>=1.9
matplotlib>=3.3.0
tqdm>=4.62.0
pandas>=1.3.0
scikit-learn>=1.0.0
plotly>=5.0.0
channels>=4.0.0
channels-redis>=4.0.0
redis>=4.0.0